backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pymongo==4.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
email-validator==2.3.0

